% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{ADT_Checker}
\alias{ADT_Checker}
\title{Analyze CITE-seq Data}
\usage{
ADT_Checker(
  object = NULL,
  RNA_assay = "RNA",
  ADT_assay = "ADT",
  umap_object = "umap",
  cell_annotations = "orig.ident"
)
}
\arguments{
\item{object}{Seurat Object}

\item{RNA_assay}{Assay name (RNA)}

\item{ADT_assay}{Assay name (ADT)}

\item{umap_object}{Name of the UMAP or TSNE object}

\item{cell_annotations}{Name of the cell annotation column from meta.data}
}
\value{
A shiny dashboard with interactive plots and export functionality
}
\description{
Analyze CITE-seq Data
}
\details{
The dashboard includes:
\itemize{
\item Interactive biplot and UMAP selection for both RNA and ADT data
\item Export selected cells as subset Seurat objects to global environment
\item Summary tables of selected cell populations
}
}
