% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{RNA_Checker}
\alias{RNA_Checker}
\title{Analyze scRNA-seq Data}
\usage{
RNA_Checker(
  object = NULL,
  RNA_assay = "RNA",
  umap_object = "umap",
  cell_annotations = "orig.ident"
)
}
\arguments{
\item{object}{Seurat Object}

\item{RNA_assay}{Assay name}

\item{umap_object}{Name of the UMAP or TSNE object}

\item{cell_annotations}{Name of the cell annotation column from meta.data}
}
\value{
A shiny dashboard with interactive plots and export functionality
}
\description{
Analyze scRNA-seq Data
}
\details{
The dashboard includes:
\itemize{
\item Interactive biplot and UMAP selection
\item Export selected cells as subset Seurat objects to global environment
\item Summary tables of selected cell populations
}
}
